Sprint 7.f

+++ CCSystem/src/main/java/com/team08/CCSystem/model/ClinicMark.java
@@ -27,7 +27,7 @@ public class ClinicMark {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(name = "mark", nullable = false, unique = false)
 	private float mark; //from 0 to 5
@@ -44,7 +44,7 @@ public class ClinicMark {
 	 * @param patient
 	 * @param clinic
 	 */
-	public ClinicMark(Integer id, float mark, Patient patient, Clinic clinic) {
+	public ClinicMark(Long id, float mark, Patient patient, Clinic clinic) {
 		super();
 		this.id = id;
 		this.mark = mark;
@@ -59,11 +59,11 @@ public class ClinicMark {
 		super();
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
